Add preliminary wayland support
I tested this with a current Ubuntu 20.04 installation running
wayland. According to strace, it's using the 'real' wayland
libs and not the xwayland wrapper.

Due to the lack of hardware, I wasn't able to test this in
some sort of kiosk environment using a Core image...

According to [1], SDL >= 2.0.4 is fully compatible with wayland.

[1] https://www.phoronix.com/scan.php?page=news_item&px=SDL-2.0.4-Released