Add drascula-music .deb file
This provides the (currently missing) background music for drascula.

Since the drascula-music package depends on the drascula package itself,
I decided to pull it in the manual way in order to avoid conflicts with
the datafile music.

If there's a way to pull in multiple .deb files via the source plugin, let me know.