Switch nightly builds to the ScummVM 2.2.0 release branch
We just started a new release cycle for ScummVM, so for the next days until the release, the 'edge' channel should reflect this branch to help finding bugs. From now on, the edge channel builds will only receive bugfixes, no new features.

I'll revert this to master after the 2.2.0 release is out.